requirements.txt
google-auth
google-cloud-datastore
pytest
requests
sh
